% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.ANOVA.assumptions.R
\name{"check.ANOVA.assumptions"}
\alias{"check.ANOVA.assumptions"}
\alias{check.ANOVA.assumptions}
\title{Check basic ANOVA assumptions}
\usage{
check.ANOVA.assumptions(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{A wide-format data frame.}

\item{alpha}{The significance level.}

\item{...}{Optional parameters.}
}
\value{
A wide-format data frame.
}
\description{
Check the normality assumption of the input dataset using the Kolmogorov-Smirnov Test and the homogeneity of variances assumption of the input dataset using the Levene's test.
}
\examples{
check.ANOVA.assumptions(example)

}
\author{
Chakkrit Tantithamthavorn (kla@chakkrit.com)
}
