% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2wide.R
\name{"long2wide"}
\alias{"long2wide"}
\alias{long2wide}
\title{Convert data from long format to wide format}
\usage{
long2wide(x, ...)
}
\arguments{
\item{x}{A long-format data frame.}

\item{...}{Optional parameters.}
}
\value{
A wide-format data frame.
}
\description{
Convert data from long format to wide format
}
\examples{
long2wide(melt(example, id.vars=0))

}
\author{
Chakkrit Tantithamthavorn (kla@chakkrit.com)
}
