% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{"normalize"}
\alias{"normalize"}
\alias{normalize}
\title{Normalize non-normal distributions using the Box-Cox Power Transformation}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{A wide-format data frame.}

\item{...}{Optional parameters.}
}
\value{
A wide-format data frame.
}
\description{
Normalize non-normal distributions using the Box-Cox Power Transformation
}
\examples{
normalized.data <- normalize(example)

}
\author{
Chakkrit Tantithamthavorn (kla@chakkrit.com)
}
