% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ScottKnottESD.R
\name{ScottKnottESD}
\alias{ScottKnottESD}
\title{The Scott-Knott Effect Size Difference (ESD) test}
\usage{
ScottKnottESD(data)
}
\arguments{
\item{data}{a long-format data frame}
}
\value{
a SK object
}
\description{
An enhancement of the Scott-Knott test (which cluster distributions into statistically distinct ranks) that takes effect size into consideration.
}
\examples{
sk <- ScottKnottESD(example)
sk$original  # Original Groups
sk$groups    # Corrected Groups with effect size wise
sk$reverse   # Reversed Groups
}

