% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SK.ESD.R
\name{SK.ESD}
\alias{SK.ESD}
\title{The Scott-Knott Effect Size Difference (ESD) test}
\usage{
SK.ESD(x, ...)
}
\arguments{
\item{data}{a long-format data frame}
}
\value{
a SK object
}
\description{
An enhancement of the Scott-Knott test (which cluster distributions into statistically distinct ranks) that takes effect size into consideration.
}
\examples{
sk <- SK.ESD(example)
sk$original  # Original Groups
sk$groups    # Corrected Groups with effect size wise
sk$reverse   # Reversed Groups

sk <- SK.ESD(melt(example), long=TRUE)

}

