% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{Example datasets of Breiman's variable importance scores}
\format{A data frame with 1,000 rows and 9 variables:
\describe{
  \item{LOC}{lines of code}
  \item{Components}{the numbers of components}
  \item{Subsystem}{the numbers of subsystems}
  \item{Files}{the numbers of files}
  \item{Commit}{the numbers of commits}
  \item{Churn}{the numbers of churns}
  \item{Ownership}{ownership}
  \item{Authorship}{authorship}
  \item{Experience}{developer's experience}
  ...
}}
\source{
\url{https://github.com/klainfo/ScottKnottESD/}
}
\usage{
example
}
\description{
A dataset containing software metrics of 1,000 calculation of Breiman's variable importance scores
}
\keyword{datasets}

