% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sk_esd.R
\name{print.sk_esd}
\alias{print.sk_esd}
\title{Print sk_esd objects}
\usage{
\method{print}{sk_esd}(x, ...)
}
\arguments{
\item{x}{A sk_esd object}

\item{...}{Optional parameters.}
}
\value{
The sk.esd ranks
}
\description{
S3 method to print sk_esd objects.
}
